\name{plot.crossvalidate}
\alias{plot.crossvalidate}
\title{
Phylopars cross-validation
}
\description{
Plots cross-validation results
}
\usage{
\method{plot}{crossvalidate}(x, \dots)
}
\arguments{
  \item{x}{
Crossvalidate object
}
  \item{\dots}{
}
}
\value{
Returns a summary of cross-validation statistics.
}
\references{
Bruggeman J, Heringa J and Brandt BW. (2009) PhyloPars: estimation of missing parameter values using phylogeny. \emph{Nucleic Acids Research} 37: W179-W184.
}
\author{
Eric W. Goolsby \email{eric.goolsby.evolution@gmail.com}
}