\name{print.crossvalidate}
\alias{print.crossvalidate}
\title{
Phylopars cross-validation
}
\description{
Generic S3 method for crossvalidate objects
}
\usage{
\method{print}{crossvalidate}(x, \dots)
}
\arguments{
  \item{x}{
Crossvalidate object
}
  \item{\dots}{
}
}
\value{
Prints cross-validation results.
}
\references{
Bruggeman J, Heringa J and Brandt BW. (2009) PhyloPars: estimation of missing parameter values using phylogeny. \emph{Nucleic Acids Research} 37: W179-W184.
}
\author{
Eric W. Goolsby \email{eric.goolsby.evolution@gmail.com}
}