\name{print.phylopars}
\alias{print.phylopars}
\title{
Print phylopars
}
\description{
Generic S3 method for phylopars
}
\usage{
\method{print}{phylopars}(x, \dots)
}\arguments{
  \item{x}{
Fitted phylopars object
}
  \item{\dots}{
  Further arguments passed to or from other methods.
}
}